feat: about page scroll links
